codestyle fix

+++ search-commons/src/main/java/no/unit/nva/search2/model/OpenSearchSwsResponse.java
@@ -21,10 +21,10 @@ public record OpenSearchSwsResponse(
     JsonNode aggregations) {
 
     public record ShardsInfo(
-        int total,
-        int successful,
-        int skipped,
-        int failed) {
+        Long total,
+        Long successful,
+        Long skipped,
+        Long failed) {
 
     }
 
@@ -33,7 +33,7 @@ public record OpenSearchSwsResponse(
         double max_score,
         List<Hit> hits) {
         public record TotalInfo(
-            int value,
+            Long value,
             String relation) {
 
         }
@@ -52,7 +52,7 @@ public record OpenSearchSwsResponse(
 
     @NotNull
     @Transient
-    public Integer getTotalSize() {
+    public Long getTotalSize() {
         return nonNull(hits)
                    ? hits.total.value
                    : 0;
